requirements.txt
taipy==3.1.1
openai==1.30.1
pillow==10.3.0
jieba
rank_bm25